Split autotester contract tests by responsibility
